requirements.txt

fastapi
uvicorn
tensorflow
pillow
numpy
requests
python-multipart
